renamed 'enum_rain_input_type.py' to 'rainfall_enum.py' - all Enums used in rainfall module are in this script now
