+++ src/dynamic_boundary_conditions/model_input.py
@@ -16,7 +16,7 @@ from geocube.api.core import make_geocube
 from src import config
 from src.digitaltwin import setup_environment
 from src.dynamic_boundary_conditions import main_rainfall, thiessen_polygons, hirds_rainfall_data_from_db, hyetograph
-from src.dynamic_boundary_conditions.enum_rain_input_type import RainInputType, HyetoMethod
+from src.dynamic_boundary_conditions.rainfall_enum import RainInputType, HyetoMethod
 
 log = logging.getLogger(__name__)
 log.setLevel(logging.DEBUG)
